Remove confirmation_token in favor of update_token.
This also solves a bug noticed in June 2019 involving subscriptions imported from Mailchimp. When these people resubscribed, we sent them a confirmation message and attempted to record that in confirmation_sent_at. Doing so however triggered an SQL constraint error because confirmation_sent_at required a confirmation_token, but that was never generated for Mailchimp imports.